+++ packages/kiarina-utils-file/src/kiarina/utils/encoding/_operations/detect_with_charset_normalizer.py
@@ -2,7 +2,7 @@ import logging
 
 from charset_normalizer import detect
 
-from ..settings import settings_manager
+from .._settings import settings_manager
 
 logger = logging.getLogger(__name__)
