+++ packages/kiarina-utils-file/src/kiarina/utils/encoding/_settings.py
