+++ packages/kiarina-utils-file/src/kiarina/utils/ext/_operations/detect_with_dictionary.py
@@ -1,6 +1,6 @@
 from .._utils.normalize_extension import normalize_extension
 from .._utils.normalize_mime_type import normalize_mime_type
-from ..settings import settings_manager
+from .._settings import settings_manager
 
 
 def detect_with_dictionary(
