+++ packages/kiarina-utils-file/src/kiarina/utils/mime/_helpers/apply_mime_alias.py
@@ -1,4 +1,4 @@
-from ..settings import settings_manager
+from .._settings import settings_manager
 
 
 def apply_mime_alias(
