+++ packages/kiarina-utils-file/src/kiarina/utils/mime/_settings.py
